Add pointInPolygon (point, p) It determines is or not the 'point' in polygon 'p'. Returns boolean in numerical representation: 0 or 1
How it work: https://goo.gl/m88H91
'The algorithm is based on the idea of counting the number of intersections of the ray coming from a given point in the direction of the horizontal axis, with the sides of the polygon. If it is parity, the point does not belong to the polygon. In this algorithm, the beam is directed to the left.'